backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic-settings>=2.7.0
aiosqlite>=0.20.0
yt-dlp>=2024.12.0
shazamio>=0.7.0
easyocr>=1.7.0
spotipy>=2.24.0
python-multipart>=0.0.18
